Fix test helper PSR-4 layout
